Revert to audio combined with video. Dedicated audio backchannel class.
