+++ cloudFrontEnd/src/app/multi-cam-view/multi-cam-view.component.scss
@@ -60,7 +60,7 @@ tr.detail-row {
 .element-detail {
    overflow: hidden;
    display: flex;
-}
+ }
 
 .section {
   margin: 12px 0;
